+++ app/src/main/java/org/glucosio/android/db/PressureReading.java
@@ -20,42 +20,42 @@
 
 package org.glucosio.android.db;
 
-import java.util.Date;
-
 import io.realm.RealmObject;
 import io.realm.annotations.PrimaryKey;
+
+import java.util.Date;
 
 public class PressureReading extends RealmObject {
     @PrimaryKey
     private long id;
 
-    private int minReading;
-    private int maxReading;
+    private double minReading;
+    private double maxReading;
     private Date created;
 
     public PressureReading() {
     }
 
-    public PressureReading(int minReading, int maxReading, Date created) {
+    public PressureReading(double minReading, double maxReading, Date created) {
         // mm/Hg
         this.minReading = minReading;
         this.maxReading = maxReading;
         this.created = created;
     }
 
-    public int getMinReading() {
+    public double getMinReading() {
         return minReading;
     }
 
-    public void setMinReading(int minReading) {
+    public void setMinReading(double minReading) {
         this.minReading = minReading;
     }
 
-    public int getMaxReading() {
+    public double getMaxReading() {
         return maxReading;
     }
 
-    public void setMaxReading(int maxReading) {
+    public void setMaxReading(double maxReading) {
         this.maxReading = maxReading;
     }
